vcからセルのリストをmodelに移行
