+++ SurroundingRestaurantsOfNPB/Controller/NPBListViewController.swift
@@ -1,5 +1,5 @@
 //
-//  MNPBListViewController.swift
+//  NPBListViewController.swift
 //  SurroundingRestaurantsOfNPB
 //
 //  Created by 金田岳大 on 2022/07/07.
@@ -8,13 +8,11 @@
 import UIKit
 
 final class NPBListViewController: UIViewController {
-    // 12球団の球場入れる場合はここで球場名を増やす
-    private let stadiumNamesList = ["ZOZOマリンスタジアム", "東京ドーム", "ベルーナドーム"]
+
     @IBOutlet private weak var tableView: UITableView! {
         didSet {
             let nib = UINib(nibName: NPBListCell.className, bundle: nil)
             tableView.register(nib, forCellReuseIdentifier: NPBListCell.className)
-            tableView.dataSource = self
             tableView.delegate = self
             tableView.separatorColor = .black
         }
@@ -22,18 +20,6 @@ final class NPBListViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-    }
-}
-
-extension NPBListViewController: UITableViewDataSource {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        stadiumNamesList.count
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: NPBListCell.className, for: indexPath) as? NPBListCell else { return UITableViewCell() }
-        cell.setupStadiumName(name: stadiumNamesList[indexPath.row])
-        return cell
     }
 }
 
@@ -46,3 +32,4 @@ extension NPBListViewController: UITableViewDelegate {
         250
     }
 }
+
